% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaspModuleBundleManager.R
\name{installJaspModuleBundle}
\alias{installJaspModuleBundle}
\title{Install Module Bundle}
\usage{
installJaspModuleBundle(installPath, bundlePath)
}
\arguments{
\item{installPath}{Path to the root of the module install folder}

\item{bundlePath}{Path to bundle to install}
}
\description{
Install Module Bundle into the relevant dirs. See JaspModuleManagement.pdf for an overview
}
