% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaspModuleBundleManager-package.R
\docType{package}
\name{jaspModuleBundleManager-package}
\alias{jaspModuleBundleManager}
\alias{jaspModuleBundleManager-package}
\title{jaspModuleBundleManager: Helper Package for Installing JASP Modules}
\description{
Creates and Manages JASP module bundles
}
\keyword{internal}
